+++ src/OSK.Petra.Inputs/Notifications/DeviceInputNotification.cs
@@ -3,9 +3,9 @@ using System;
 
 namespace OSK.Petra.Inputs.Notifications;
 
-public class DeviceInputNotification(RuntimeDeviceIdentifier deviceIdentifier, int inputId, TimeSpan deltaTime, params IInputEvent[] inputEvents): DeviceNotification(deviceIdentifier)
+public class DeviceInputNotification(RuntimeDeviceIdentifier deviceIdentifier, long inputId, TimeSpan deltaTime, params IInputEvent[] inputEvents): DeviceNotification(deviceIdentifier)
 {
-    public int InputId => inputId;
+    public long InputId => inputId;
 
     public TimeSpan DeltaTime => deltaTime;
